% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp.R
\name{calculateBindingCharacteristics}
\alias{calculateBindingCharacteristics}
\title{Get binding characteristics}
\usage{
calculateBindingCharacteristics(IP, input, read.len = NULL,
  min.shift = -500, max.shift = 1500, bin.size = 5, cluster = NULL)
}
\arguments{
\item{IP}{File name/GAlignments object/\code{spp} tags of IP sample.}

\item{input}{File name/GAlignments object/\code{spp} tags of input sample.}

\item{read.len}{Read length. If \code{NULL}, the read length is detected automatically by using the median of the first 10,000 reads in the first ChIP sample.}

\item{min.shift}{Minimum shift range.}

\item{max.shift}{Maximum shift range.}

\item{bin.size}{Bin tags within the specified number of basepairs to speed up calculation.}

\item{cluster}{A \code{snow} cluster: \code{cluster <- snow::makeCluster(cores)}.}
}
\value{
A list with the ChIP data, the input data and the binding characteristics.
}
\description{
Get binding characteristics from cross-correlation profile using the
\code{spp} package.
}
\details{
Types of analysis:
\itemize{
\item Calculate binding characteristics.
\item Report QC metrics, RSC and NSC.
\item Calculate genome-wide profiles of smoothed tag density.
\item Calculate genome-wide profiles providing conservative statistical estimates of fold enrichment.
\item Determine statistically significant point binding positions.
\item Assessing saturation properties.
}
}
\examples{
# call SPP to get the binding characteristics and plot it using phantomPeak()
bc <- calculateBindingCharacteristics("IP.bam", "input.bam", read.len=51)
pp <- phantomPeak(bc$chip.data, bc$input.data, bc$bc)
}

