% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idr.R
\name{calculateIDR}
\alias{calculateIDR}
\title{Calculate irreproducibly discovery rate}
\usage{
calculateIDR(chipSamples, inputSamples, org, assembly, version,
  readLength = NULL, shiftRange = c(-500, 1500), binSize = 5,
  crossCorrelationPeakShift = NULL, invalidCrossCorrelationPeaks = c(10,
  readLength + 10), halfWidth = NULL, overlapRatio = 0, isBroadPeak = F,
  cluster = NULL)
}
\arguments{
\item{chipSamples}{vector of file names of immuno-precipitated samples in \code{.bam} format or list of \code{GAlignments} objects}

\item{inputSamples}{vector of file names of control samples in \code{.bam} format or list of \code{GAlignments} objects}

\item{org}{organism as described in the \code{BSGenome} package: \code{Hsapiens}, \code{Mmusculus}, ...}

\item{assembly}{assembly name: \code{UCSC}, \code{NCBI}, \code{TAIR}, ...}

\item{version}{assembly version: \code{hg19}, \code{mm9}, ...}

\item{readLength}{length of reads (helps identify phantom peak), if \code{NULL} then it is determined automatically from the first 10000 reads in the BAM file}

\item{shiftRange}{strand shifts at which cross-correlation is evaluated}

\item{binSize}{step size for \code{shiftRange}}

\item{crossCorrelationPeakShift}{user-defined cross-correlation peak shift (when given, SPP does not try to detect the peak shift automatically)}

\item{invalidCrossCorrelationPeaks}{strand shifts to exclude (to avoid phantom peaks)}

\item{halfWidth}{a numerical value to truncate the peaks to; \code{NULL} means use peak width reported by SPP}

\item{overlapRatio}{a value between 0 and 1. It controls how much overlaps two peaks need to have to be considered as calling the same region. It is the ratio of overlap / short peak of the two. When set to 0, two peaks are deemed as calling the same region, if they overlap by at least 1 bp}

\item{isBroadPeak}{if broadpeak is used, set to \code{T}; if narrowpeak is used, set to \code{F}}

\item{cluster}{a \code{snow} cluster: \code{cluster <- snow::makeCluster(4)}}
}
\value{
object that is suitable as input to \code{plotIDR}
}
\description{
Assess the consistency between ChIP-Seq replicates according to Qunhua Li et al.
}
\details{
Assess the consistency between ChIP-Seq replicates according to Qunhua Li et al.:

Ref: Qunhua Li, James B. Brown, Haiyan Huang, and Peter J. Bickel: Measuring reproducibility of high-throughput experiments.
     Ann Appl Stat. 2011 October 13

The irreproducible discovery rate (IDR) is a measure for the probability
that a peak is called in a ChIP-Seq sample, if it has been called in another
sample. Peaks that result from biological activity should be called
consistently between replicates. They are assigned a low IDR. In contrast,
peaks that are noise are typically not called in all replicates and are
assigned a high IDR.
The output of this function is meant to be used as input to the function
\code{plotIDR}, which generates plots for the assessment of the IDR of
ChIP-Seq replicates. Refer to Qunhua Li et al (2011) for an explanation on
how to interpret the plots.
}
\examples{
chipTuples <- calculateIDR(c("IP1.bam", "IP2.bam"), c("input1.bam", "input2.bam"),
                            "Hsapiens", "UCSC", "hg19")
for (chipTuple in chipTuples) {
    pdf(paste(basename(chipTuple$rep1), "_VS_", basename(chipTuple$rep2), ".pdf", sep=""), paper="a4r", width=11, height=8.5)
    plotIDR(chipTuple)
    dev.off()
}
}

