% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp.R
\name{callPeaks}
\alias{callPeaks}
\title{Call peaks using SPP}
\usage{
callPeaks(chip.data, input.data, binding.characteristics, fdr = 0.01,
  method = tag.wtd, window.size = 1000, z.thr = 3, cluster = NULL)
}
\arguments{
\item{chip.data}{\code{spp} tags of the ChIP sample calculated with \code{calculateBindingCharacteristics}.}

\item{input.data}{\code{spp} tags of the input sample calculated with \code{calculateBindingCharacteristics}.}

\item{binding.characteristics}{The binding characteristics calculated with \code{calculateBindingCharacteristics}.}

\item{fdr}{Peaks with a corrected p-value above this threshold are not reported.}

\item{method}{The peak calling method, either \code{tag.wtd} or \code{tag.lwcc}.}

\item{window.size}{The window size to be used in calculating enrichment.}

\item{z.thr}{A z-score corresponding to the Poisson ratio threshold used to flag significantly enriched windows.}

\item{cluster}{A snow cluster: \code{cluster <- snow::makeCluster(cores)}.}
}
\value{
\code{spp} peak data as generated by \code{find.binding.positions} and \code{add.broad.peak.regions}.
}
\description{
Call peaks using the \code{spp} package.
}
\details{
\code{callPeaks()} uses the \code{spp} package to identify regions enriched
with reads (peaks) when comparing a ChIP sample to a control sample.
}
\examples{
bc <- calculateBindingCharacteristics("IP.bam", "input.bam", read.len=51)
peaks <- callPeaks(bc$chip.data, bc$input.data, bc$bc)
}

