% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp.R
\name{frip}
\alias{frip}
\title{Fraction of reads in peaks}
\usage{
frip(reads, peaks, singleEnd = T)
}
\arguments{
\item{reads}{File path, \code{GRanges}, \code{GAlignments}, \code{GAlignmentPairs}, \code{BamViews} or \code{BamFile} object containing reads from ChIP sample.}

\item{peaks}{\code{GRanges} object indicating positions and widths of peaks or \code{spp} peaks as output by \code{find.binding.positions} or \code{add.broad.peak.regions}.}

\item{singleEnd}{Only applicable when \code{reads} is a \code{BamViews} or \code{BamFile} object; set to \code{TRUE} for single-end data, \code{FALSE} for paired-end data.}
}
\value{
Fraction of reads in peaks.
}
\description{
Calculate the FRiP metric (fraction of reads in peaks).
}
\details{
\code{frip} counts the number of mapped reads that fall into regions
which were identified as binding positions (peaks) in a ChIP-Seq experiment.
It returns the ratio [mapped reads in peaks] / [all mapped reads]. High
values are an indicator of a good signal-to-noise ratio.

Note that the method used for peak identification influences the FRiP value.
A comparison of FRiP values between experiments is only possible, if the
same method was used for identifying peaks (i.e., same peak-caller with
identical parameters).

Ref: Stephen G. Landt, Georgi K. Marinov, Anshul Kundaje, Pouya Kheradpour,
Florencia Pauli, Serafim Batzoglou, Bradley E. Bernstein, Peter Bickel,
James B. Brown, Philip Cayting, Yiwen Chen, Gilberto DeSalvo,
Charles Epstein, Katherine I. Fisher-Aylor, Ghia Euskirchen, Mark Gerstein,
Jason Gertz, Alexander J. Hartemink, Michael M. Hoffman,
Vishwanath R. Iyer, Youngsook L. Jung, Subhradip Karmakar, Manolis Kellis,
Peter V. Kharchenko, Qunhua Li, Tao Liu, X. Shirley Liu, Lijia Ma,
Aleksandar Milosavljevic, Richard M. Myers, Peter J. Park,
Michael J. Pazin, Marc D. Perry, Debasish Raha, Timothy E. Reddy,
Joel Rozowsky, Noam Shoresh, Arend Sidow, Matthew Slattery,
John A. Stamatoyannopoulos, Michael Y. Tolstorukov, Kevin P. White,
Simon Xi, Peggy J. Farnham, Jason D. Lieb, Barbara J. Wold, Michael Snyder:
ChIP-seq guidelines and practices of the ENCODE and modENCODE consortia
Genome Res. Sep 2012; 22(9): 1813–1831.
}
\examples{
library(rtracklayer)
frip("chip.bam", import.bed("peaks.bed"))
}

