% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp.R
\name{gAlignmentsToSppTags}
\alias{gAlignmentsToSppTags}
\title{Convert GAlignments object to SPP tags}
\usage{
gAlignmentsToSppTags(reads)
}
\arguments{
\item{reads}{\code{GAlignments} object to converted to \code{spp} tags.}
}
\value{
\code{spp} tags which are compatible with the output of \code{read.bam.tags}
}
\description{
\code{gAlignmentsToSppTags} converts a \code{GAlignments} object to
\code{spp} tags, for the purpose of passing \code{GAlignments} objects
to functions of the \code{spp} package.
}
\details{
\code{spp} uses a custom format to represent tags (see \code{read.bam.tags}).
This helper function converts \code{GAlignments} objects to this custom
format. The \code{GAlignments} class is more versatile than the custom tag
format of \code{spp}. If several metrics shall be calculated on the same BAM
file, then it makes sense to load it only once into memory as a
\code{GAlignments} object and convert it to \code{spp} tags for
\code{spp}-related metrics, instead of loading the BAM file again just to
have it in \code{spp} tag format. Note that the \code{GAlignments} object
should contain the \code{NM} flag (number of mismatches). Otherwise, the
\code{quality} attribute of \code{spp} tags cannot be populated with values.
The \code{spp} function \code{select.informative.tags} simply accepts all
tags, when the \code{quality} attribute is missing. Create the
\code{GAlignments} object using

\code{readGAlignments(..., param=ScanBamParam(tag="NM"))}

to include the number of mismatches in the \code{GAlignments} object.
}
\examples{
# the following three lines essentially perform the same as read.bam.tags()
# but a GAlignments object is more versatile and serves as input to many
# functions
library(GenomicAlignments)
chipSample <- readGAlignments("IP.bam", param=ScanBamParam(tag="NM"))
sppTags <- gAlignmentsToSppTags(chipSample)
}

