% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idr.R
\name{getCrossConsistencyRatios}
\alias{getCrossConsistencyRatios}
\title{Get cross-consistency ratios}
\usage{
getCrossConsistencyRatios(chipTuples, idrCutoff = 0.01)
}
\arguments{
\item{chipTuples}{output of \code{calculateIDR}}

\item{idrCutoff}{count only peaks below this IDR value (for mammalian cells a value of 0.01 is recommended)}
}
\value{
list of quotients of peak counts (pooled pseudo-replicates / true replicates)
}
\description{
Calculate the ratio of peaks below a given IDR threshold between all true replicates and the pooled pseudo-replicates
}
\details{
The ChIP-Seq guidelines of the ENCODE consortium recommend that the number
of peaks with an IDR below 0.01 should not differ by more than a factor of 2
between the pair of pooled pseudo-replicates and any pair of true replicates.
This function counts the peaks of all pairs of true replicates with an IDR
below 0.01 as well as the peaks of pseudo-replicates from a pooled sample.
The output is a list of quotients calculated as:

[peak count of pooled replicates] / [peak count of true replicate]

All values should be lower than or equal to 2. Higher values indicate low
reproducibility.

Ref: Stephen G. Landt, Georgi K. Marinov, Anshul Kundaje, Pouya Kheradpour,
Florencia Pauli, Serafim Batzoglou, Bradley E. Bernstein, Peter Bickel,
James B. Brown, Philip Cayting, Yiwen Chen, Gilberto DeSalvo,
Charles Epstein, Katherine I. Fisher-Aylor, Ghia Euskirchen, Mark Gerstein,
Jason Gertz, Alexander J. Hartemink, Michael M. Hoffman,
Vishwanath R. Iyer, Youngsook L. Jung, Subhradip Karmakar, Manolis Kellis,
Peter V. Kharchenko, Qunhua Li, Tao Liu, X. Shirley Liu, Lijia Ma,
Aleksandar Milosavljevic, Richard M. Myers, Peter J. Park,
Michael J. Pazin, Marc D. Perry, Debasish Raha, Timothy E. Reddy,
Joel Rozowsky, Noam Shoresh, Arend Sidow, Matthew Slattery,
John A. Stamatoyannopoulos, Michael Y. Tolstorukov, Kevin P. White,
Simon Xi, Peggy J. Farnham, Jason D. Lieb, Barbara J. Wold, Michael Snyder:
ChIP-seq guidelines and practices of the ENCODE and modENCODE consortia
Genome Res. Sep 2012; 22(9): 1813–1831.
}
\examples{
chipTuples <- calculateIDR(c("IP1.bam", "IP2.bam"), c("input1.bam", "input2.bam"),
                            "Hsapiens", "UCSC", "hg19")
getCrossConsistencyRatios(chipTuples)
}

