% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IPstrength.R
\name{IPstrength}
\alias{IPstrength}
\title{Plot IP strength}
\usage{
IPstrength(IP, input)
}
\arguments{
\item{IP}{tag counts of ChIP sample as produced by \code{tagCount}.}

\item{input}{tag counts of input sample as produced by \code{tagCount}.}
}
\value{
The two metrics, alpha and enrichment.
}
\description{
Plot IP strength in the way described in the CHANCE paper.
}
\details{
Plot IP strength in the way described in the CHANCE paper:

Ref: Diaz, A et al. Normalization, bias correction, and peak calling for ChIP-seq.
     Stat Appl Genet Mol Biol. 2012 March 31

The process follows these steps:
\enumerate{
\item Tile the genome into bins and count the tags on each bin.
\item Sort bins of IP sample and sort input in the same order.
\item Calculate for IP and input the cumulative sum of reads.
\item Calculate for IP and input the cumulative percentages.
\item Calculate the alpha scaling factor (ratio of cumulative sums between
      IP and input at the bin with the highest difference).
\item Plot IP strength.
}
}
\examples{
counts <- tagCount(
    samples=c("IP.bam", "input.bam"),
    org="Hsapiens", assembly="UCSC", version="hg19"
)
IPstrength(counts["IP.bam"], counts["input.bam"])
}

