% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp.R
\name{mser}
\alias{mser}
\title{Assess saturation properties}
\usage{
mser(chip.data, input.data, binding.characteristics, step.size = 1e+05,
  test.agreement = 0.99, n.chains = 8, n.steps = 6, fold.enrichment = 2,
  fdr = 0.01, cluster = NULL)
}
\arguments{
\item{chip.data}{\code{spp} tags of the ChIP sample calculated with \code{calculateBindingCharacteristics}.}

\item{input.data}{\code{spp} tags of the input sample calculated \code{calculateBindingCharacteristics}.}

\item{binding.characteristics}{The binding characteristics calculated with \code{calculateBindingCharacteristics}.}

\item{step.size}{The amount of tags to add in every step to reach the saturation criteria.}

\item{test.agreement}{Saturation criteria consistency in the set of binding positions when adding \code{step.size} tags.}

\item{n.chains}{The number of subsampled chains.}

\item{n.steps}{The number of steps to convergence.}

\item{fold.enrichment}{The target fold enrichment.}

\item{fdr}{FDR cutoff.}

\item{cluster}{A snow cluster: \code{cluster <- snow::makeCluster(cores)}.}
}
\value{
A vector containing the minimal saturated enrichment ratio (MSER) and the predicted sequencing depth (PSD) in million tags.
}
\description{
Determine the minimal saturated enrichment ratio (MSER) and the predicted
sequencing depth (PSD).
}
\details{
Determine the minimal saturated enrichment ratio (MSER). The saturation
criteria here is 99\% consistency in the set of binding positions when adding
10,000 tags.

Determine the predicted sequencing depth.
}
\examples{
# call SPP to get the binding characteristics
bc <- calculateBindingCharacteristics("IP.bam", "input.bam", read.len=51)
# calculate MSER and PSD
minser <- mser(bc$chip.data, bc$input.data, bc$bc)
}

