% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binned_pca.R
\name{normalizeTagCount}
\alias{normalizeTagCount}
\title{Normalize tag counts}
\usage{
normalizeTagCount(counts, n = 500)
}
\arguments{
\item{counts}{The output of \code{tagCount}.}

\item{n}{Limit the output to the top \code{n} bins with the greatest variance. The default value of 500 is usually sufficient for clustering. \code{NULL} means all bins are returned.}
}
\value{
A \code{data.frame} with normalized counts on bins. Each column contains the counts of one sample.
}
\description{
Normalize tag counts produced by \code{tagCount} by total tag count.
}
\details{
This function normalizes the tag counts produced by \code{tagCount}.
The tag counts per bin are scaled, such that each sample has as many
normalized tags as the sample with the most tags.
}
\examples{
counts <- tagCount(
    samples=c("ctl1.bam", "ctl2.bam", "chip1.bam", "chip2.bam"),
    org="Mmusculus", assembly="UCSC", version="mm9"
)
nCounts <- normalizeTagCount(counts)
topHeatmap(nCounts)
samplesHeatmap(nCounts)
pcaPlot(nCounts, as.factor(c("ctl", "ctl", "chip", "chip")))
}

