% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binned_pca.R
\name{pcaPlot}
\alias{pcaPlot}
\title{PCA plot of ChIP and input samples}
\usage{
pcaPlot(normalizedCounts, condition = as.factor(colnames(counts)))
}
\arguments{
\item{normalizedCounts}{The normalized \code{data.frame} with counts per bin as generated by \code{normalizeTagCount}.}

\item{condition}{A factor containing the groups which the samples belong to. Samples are colored by condition.}
}
\value{
The result of the PCA analysis.
}
\description{
Plot the first two components of the PCA analysis on all samples.
}
\details{
\code{pcaPlot} runs a PCA analysis on all samples (ChIP and input) and plots
the first two components of the analysis in a scatter plot. Each dot in the
plot represents a sample. Samples which the analysis deems closely related
cluster together.
}
\examples{
counts <- tagCount(
    samples=c("ctl1.bam", "ctl2.bam", "chip1.bam", "chip2.bam"),
    org="Mmusculus", assembly="UCSC", version="mm9"
)
nCounts <- normalizeTagCount(counts)
pcaPlot(nCounts, as.factor(c("ctl", "ctl", "chip", "chip")))
}

