% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp.R
\name{phantomPeak}
\alias{phantomPeak}
\title{QC metrics RSC and NSC}
\usage{
phantomPeak(chip.data, input.data, binding.characteristics,
  smoothing.bandwidth = 5)
}
\arguments{
\item{chip.data}{\code{spp} tags of the ChIP sample calculated with \code{calculateBindingCharacteristics}.}

\item{input.data}{\code{spp} tags of the input sample calculated \code{calculateBindingCharacteristics}.}

\item{binding.characteristics}{The binding characteristics calculated with \code{calculateBindingCharacteristics}.}

\item{smoothing.bandwidth}{Smooth the curve of the cross-correlation plot using a moving average (higher value means more smoothing).}
}
\value{
A vector containing the NSC, RSC, estimated read length and estimated fragment length.
}
\description{
Report the QC metrics, RSC and NSC, and plot the cross-correlation profile.
}
\details{
Calculate RSC and NSC metrics and plot the cross-correlation profile.
\itemize{
\item The relative strand correlation (RSC) is the ratio between the
peak of the fragment length and the peak of the read length. ENCODE
guidelines recommend that the RSC be > 0.8.
\item The normalized strand coefficient, NSC, is the normalized ratio
between the peak of the fragment-length cross-correlation and the
background cross-correlation. ENCODE guidelines recommend that the NSC be
> 1.05.
}
}
\examples{
# call SPP to get the binding characteristics
bc <- calculateBindingCharacteristics("IP.bam", "input.bam", read.len=51)
# plot the cross-correlation profile
pp <- phantomPeak(bc$chip.data, bc$input.data, bc$bc)
}

