% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idr.R
\name{plotIDR}
\alias{plotIDR}
\title{plot irreproducible discovery rate}
\usage{
plotIDR(chipTuple, idrCutoff = 0.01)
}
\arguments{
\item{chipTuple}{one item of the return value of calculateIDR}

\item{idrCutoff}{peaks with an IDR greater than this cut-off are colored red in the scatter plot}
}
\description{
Make plots from the output of the function \code{calculateIDR}
}
\details{
The output of \code{calculateIDR()} is a list of tuples of ChIP samples.
\code{plotIDR()} draws plots which compare the two members of a tuple.
For every tuple six plots are generated:
- number of peaks in common as a function of the number of _all_ significant peaks
- slope of the previous plot
- number of peaks in common as a function of the number of _matched_ significant peaks
- slope of the previous plot
- IDR as a function of the number of significant peaks
- scatter plot of ranked peaks of the ChIP samples
Refer to Qunhua Li et al (2011) for an explanation on how to interpret
the plots.

Ref: Qunhua Li, James B. Brown, Haiyan Huang, and Peter J. Bickel: Measuring reproducibility of high-throughput experiments.
     Ann Appl Stat. 2011 October 13
}
\examples{
chipTuples <- calculateIDR(c("IP1.bam", "IP2.bam"), c("input1.bam", "input2.bam"),
                            "Hsapiens", "UCSC", "hg19")
for (chipTuple in chipTuples) {
    pdf(paste(basename(chipTuple$rep1), "_VS_", basename(chipTuple$rep2), ".pdf", sep=""), paper="a4r", width=11, height=8.5)
    plotIDR(chipTuple)
    dev.off()
}
}

