% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binned_pca.R
\name{samplesHeatmap}
\alias{samplesHeatmap}
\title{Heatmap and cluster sample-sample distances}
\usage{
samplesHeatmap(normalizedCounts, ...)
}
\arguments{
\item{normalizedCounts}{The normalized \code{data.frame} with counts per bin as generated by \code{normalizeTagCount}.}

\item{...}{Extra params sent to \code{heatmap.2}.}
}
\value{
The plotted distance matrix.
}
\description{
Plot a \code{gplots} heatmap with the sample-sample distances.
}
\details{
This function uses the \code{heatmap.2} function of \code{gplots} to assess
the relationship between samples.
}
\examples{
counts <- tagCount(
    samples=c("ctl1.bam", "ctl2.bam", "chip1.bam", "chip2.bam"),
    org="Mmusculus", assembly="UCSC", version="mm9"
)
nCounts <- normalizeTagCount(counts)
samplesHeatmap(nCounts)
}

