% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp.R
\name{sppPeaksToGRanges}
\alias{sppPeaksToGRanges}
\title{Convert output from SPP to GRanges object}
\usage{
sppPeaksToGRanges(peaks)
}
\arguments{
\item{peaks}{Output of \code{find.binding.positions()} from the \code{spp} package.}
}
\value{
\code{spp} data converted to \code{GRanges} object.
}
\description{
Convert the peaks called by \code{spp} to a \code{GRanges} object.
}
\details{
\code{sppPeaksToGRanges()} takes the output of
\code{find.binding.positions()} from the \code{spp} package and converts it
into a \code{GRanges} object.
}
\examples{
bc <- calculateBindingCharacteristics("IP.bam", "input.bam")
peaks <- callPeaks(bc$chip.data, bc$input.data, bc$bc)
sppPeaksToGRanges(peaks)
}

