% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binned_pca.R
\name{tagCount}
\alias{tagCount}
\title{ChIP-seq binned tag counting tool}
\usage{
tagCount(samples, org, assembly, version, binSize = 1000, cluster = NULL)
}
\arguments{
\item{samples}{List/vector of paths or list of GAlignments/BamFile/BamViews objects or GAlignmentsList or BamFileList that will be processed.}

\item{org}{The organism as described in the BSGenome package: Hsapiens, Mmusculus, ...}

\item{assembly}{The assembly name: UCSC, NCBI, TAIR, ...}

\item{version}{The assembly version: hg19, mm9, ...}

\item{binSize}{The size of the bins.}

\item{cluster}{A \code{snow} cluster to process samples in parallel: \code{cluster <- snow::makeCluster(cores)}.}
}
\value{
A \code{data.frame} with counts on bins. Each column contains the counts of one sample.
}
\description{
\code{tagCount} returns a \code{data.frame} of tag counts per bin per sample.
}
\details{
This function makes use of the \code{GenomicAlignments} package to count
tags on bins. The reference genome is split into bins and a
\code{data.frame} with the counts per bin is returned. The \code{data.frame}
has one column for each sample.
}
\examples{
counts <- tagCount(
    samples=c("ctl1.bam", "ctl2.bam", "chip1.bam", "chip2.bam"),
    org="Mmusculus", assembly="UCSC", version="mm9"
)
nCounts <- normalizeTagCount(counts)
topHeatmap(nCounts)
samplesHeatmap(nCounts)
pcaPlot(nCounts, as.factor(c("ctl", "ctl", "chip", "chip")))
}

